refactor: reorganize code into modular structure with DTOs and shared utilities
feat(products): add products DTO mapping functions
feat(collections): add collections DTO mapping functions
feat(ai): move AI-related functions to dedicated module
feat(client): extract store info fetching logic to client module
refactor(utils): consolidate shared utility functions
refactor(tests): update test imports to reflect new module structure
